+++ src/models/Transaction.java
@@ -11,10 +11,10 @@ public class Transaction {
     private final UUID id;
     private final UUID userId;
     private final TYPE type;
-    private final long amount;
+    private final double amount;
     private ZonedDateTime timestamp;
 
-    public Transaction(UUID id, UUID userId, long amount, Transaction.TYPE type, ZonedDateTime timestamp) {
+    public Transaction(UUID id, UUID userId, double amount, Transaction.TYPE type, ZonedDateTime timestamp) {
         this.amount = amount;
         this.id = id;
         this.userId = userId;
@@ -25,7 +25,7 @@ public class Transaction {
     @Override
     public String toString() {
         return String.format(
-                "Transaction | ID: %s | User: %s | Type: %s | Amount: %d | Time: %s",
+                "Transaction | ID: %s | User: %s | Type: %s | Amount: %.2f | Time: %s",
                 id, userId, type, amount, timestamp
         );
     }
